+++ icpc-backend/src/time/time.service.ts
@@ -1,15 +1,28 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateTimeDto } from './dto/create-time.dto';
 import { UpdateTimeDto } from './dto/update-time.dto';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository, createQueryBuilder } from 'typeorm';
+import { Repository } from 'typeorm';
 import { Time } from './entities/time.entity';
+import { Comment } from 'src/comment/entities/comment.entity';
+import {
+  Ticket,
+  TicketOperation,
+  TicketStatus
+} from 'src/ticket/entities/ticket.entity';
+import { Excercise } from 'src/excercises/entities/excercise.entity';
 
 @Injectable()
 export class TimeService {
   constructor(
     @InjectRepository(Time)
-    private readonly timeRepository: Repository<Time>
+    private readonly timeRepository: Repository<Time>,
+    @InjectRepository(Comment)
+    private readonly commentRepository: Repository<Comment>,
+    @InjectRepository(Ticket)
+    private readonly ticketRepository: Repository<Ticket>,
+    @InjectRepository(Excercise)
+    private readonly excerciseRepository: Repository<Excercise>
   ) {}
 
   async create(createTimeDto: CreateTimeDto) {
@@ -50,7 +63,41 @@ export class TimeService {
   }
 
   async remove(id: string) {
-    const time = await this.timeRepository.findOneBy({ id });
-    return await this.timeRepository.remove(time);
+    const allTimeLimits = await this.timeRepository.find();
+    if (allTimeLimits.length === 1) {
+      throw new BadRequestException(
+        'No se puede eliminar el único límite de tiempo'
+      );
+    }
+    let pivot = allTimeLimits[0];
+    if (pivot.id === id) {
+      pivot = allTimeLimits[1];
+    }
+    const time = await this.timeRepository
+      .createQueryBuilder('time')
+      .where('time.id = :id', { id })
+      .leftJoinAndSelect('time.excercises', 'excercises')
+      .getOne();
+    const ticketComment = `El límite de tiempo ${time.timeLimit.toString()} fue eliminado`;
+    const commmentId = this.commentRepository.create({
+      body: ticketComment
+    });
+    const savedComment = await this.commentRepository.save(commmentId);
+    const ticket = this.ticketRepository.create({
+      operation: TicketOperation.DELETE,
+      status: TicketStatus.ACCEPTED,
+      commentId: savedComment
+    });
+    if (ticket) {
+      if (time.excercises.length > 0) {
+        for (const excercise of time.excercises) {
+          excercise.time = pivot;
+          await this.excerciseRepository.save(excercise);
+        }
+      }
+      return await this.timeRepository.remove(time);
+    } else {
+      throw new BadRequestException('Error al eliminar el límite de tiempo');
+    }
   }
 }
